src/main/java/advprog/bot/feature/xkcd/XkcdChatHandler.java
package advprog.bot.feature.xkcd;

import advprog.bot.feature.xkcd.helper.Xkcd;
import advprog.bot.line.AbstractLineChatHandlerDecorator;
import advprog.bot.line.LineChatHandler;

import com.linecorp.bot.model.event.MessageEvent;
import com.linecorp.bot.model.event.message.AudioMessageContent;
import com.linecorp.bot.model.event.message.ImageMessageContent;
import com.linecorp.bot.model.event.message.LocationMessageContent;
import com.linecorp.bot.model.event.message.StickerMessageContent;
import com.linecorp.bot.model.event.message.TextMessageContent;
import com.linecorp.bot.model.message.ImageMessage;
import com.linecorp.bot.model.message.Message;
import com.linecorp.bot.model.message.TextMessage;

import java.util.Collections;
import java.util.List;
import java.util.logging.Logger;

import org.springframework.stereotype.Service;

// this is for example purpose !!!
public class XkcdChatHandler extends AbstractLineChatHandlerDecorator {

    private static final Logger LOGGER = Logger.getLogger(XkcdChatHandler.class.getName());

    public XkcdChatHandler(LineChatHandler decoratedHandler) {
        this.decoratedLineChatHandler = decoratedHandler;
        LOGGER.info("Xkcd chat handler added!");
    }

    @Override
    protected boolean canHandleTextMessage(MessageEvent<TextMessageContent> event) {
        return event.getMessage().getText().split(" ")[0].equals("/xkcd");
    }

    @Override
    protected List<Message> handleTextMessage(MessageEvent<TextMessageContent> event) {
        String[] pesan = event.getMessage().toString().split(" ");
        if (pesan[0].equals("/xkcd")) {
            String imageUrl = Xkcd.gambar(event.getMessage().getText());
            if (! imageUrl.endsWith("!")) {
                return Collections.singletonList(
                        new ImageMessage(imageUrl,imageUrl)
                );
            }
            return Collections.singletonList(
                    new TextMessage("Gambar salah atau format tidak ada")
            );
        }
        return Collections.singletonList(
                new TextMessage("format salah")
        );

    }

    @Override
    protected boolean canHandleImageMessage(MessageEvent<ImageMessageContent> event) {
        return false;
    }

    @Override
    protected boolean canHandleAudioMessage(MessageEvent<AudioMessageContent> event) {
        return false;
    }

    @Override
    protected boolean canHandleStickerMessage(MessageEvent<StickerMessageContent> event) {
        return false;
    }

    @Override
    protected boolean canHandleLocationMessage(MessageEvent<LocationMessageContent> event) {
        return false;
    }
}
